DM and notification views
