+++ bookwyrm/views/__init__.py
@@ -3,3 +3,6 @@ from .authentication import Login, Register, Logout
 from .password import PasswordResetRequest, PasswordReset, ChangePassword
 from .invite import ManageInvites, Invite
 from .landing import About, Home, Feed, Discover
+from .notifications import Notifications
+from .direct_message import DirectMessage
+from .import_datra import Import
