+++ bookwyrm/vviews.py
@@ -64,31 +64,6 @@ def not_found_page(request, _):
         request, 'notfound.html', {'title': 'Not found'}, status=404)
 
 
-@login_required
-@require_GET
-def direct_messages_page(request, page=1):
-    ''' like a feed but for dms only '''
-    activities = get_activity_feed(request.user, 'direct')
-    paginated = Paginator(activities, PAGE_LENGTH)
-    activity_page = paginated.page(page)
-
-    prev_page = next_page = None
-    if activity_page.has_next():
-        next_page = '/direct-message/?page=%d#feed' % \
-                activity_page.next_page_number()
-    if activity_page.has_previous():
-        prev_page = '/direct-messages/?page=%d#feed' % \
-                activity_page.previous_page_number()
-    data = {
-        'title': 'Direct Messages',
-        'user': request.user,
-        'activities': activity_page.object_list,
-        'next': next_page,
-        'prev': prev_page,
-    }
-    return TemplateResponse(request, 'direct_messages.html', data)
-
-
 def get_activity_feed(
         user, privacy, local_only=False, following_only=False,
         queryset=models.Status.objects):
@@ -214,22 +189,6 @@ def import_status(request, job_id):
         'failed_items': failed_items,
         'task': task
     })
-
-
-@login_required
-@require_GET
-def notifications_page(request):
-    ''' list notitications '''
-    notifications = request.user.notification_set.all() \
-            .order_by('-created_date')
-    unread = [n.id for n in notifications.filter(read=False)]
-    data = {
-        'title': 'Notifications',
-        'notifications': notifications,
-        'unread': unread,
-    }
-    notifications.update(read=True)
-    return TemplateResponse(request, 'notifications.html', data)
 
 
 @csrf_exempt
